+++ src/app/components/__tests__/SearchForm.js
@@ -4,7 +4,7 @@ import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { shallow, mount, render } from 'enzyme';
 import createComponentWithIntl from '../../../test-util/create-component-with-intl';
-import SearchForm from '../SearchForm';
+import SearchForm from '../SearchForm/index';
 
 describe('SearchFrom', () => {
   beforeEach(() => {
